give user permission to update or delete only dumplings they created

+++ dumplings/views.py
@@ -92,15 +92,22 @@ def dumpling_detail(request, id,):
   if request.method == 'GET':
     serializer = DumplingSerializer(dumpling)
     return Response(serializer.data)
-  elif request.method == 'PUT':
-    serializer = DumplingSerializer(dumpling, data=request.data)
-    if serializer.is_valid():
-      serializer.save()
-      return Response(serializer.data)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-  elif request.method == 'DELETE':
-    dumpling.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+  # check logged in user is owner
+  elif request.method in ['PUT', 'DELETE']:
+    print('Dumpling owner:', dumpling.owner)
+    print('Request user:', request.user)
+   
+    if dumpling.owner != request.user:
+      return Response({'message': 'You do not have permission to edit or delete this dumpling.'}, status=status.HTTP_403_FORBIDDEN)
+    if request.method == 'PUT':
+      serializer = DumplingSerializer(dumpling, data=request.data)
+      if serializer.is_valid():
+        serializer.save()
+        return Response(serializer.data)
+      return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    elif request.method == 'DELETE':
+      dumpling.delete()
+      return Response(status=status.HTTP_204_NO_CONTENT)
 
 #tags/
 @api_view(['GET', 'POST'])
